azure_function/tuya_proxy/requirements.txt
azure-functions
requests
pycryptodome
